+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 17.7.0 Fixes
 
 - `[Spinbox]` Updated type for step and fixed increase/decrease buttons not getting disabled when value passed with ngControl was matching min/max value. ([EP#1680](https://github.com/infor-design/enterprise-ng/issues/1680))
+- `[Dropdown]` Fix setFocus method to work also when there is no label associated with the dropdown. ([EP#1673](https://github.com/infor-design/enterprise-ng/issues/1673))
 
 ## 17.6.0